Move methods to api/graphs
